// frontend/src/types/EntertainerDetails.ts
export interface EntertainerDetails {
    entertainerId: number;
    entStageName: string | null;
    entSsn: string | null;
    entStreetAddress: string | null;
    entCity: string | null;
    entState: string | null;
    entZipCode: string | null;
    entPhoneNumber: string | null;
    entWebPage: string | null;
    entEmailAddress: string | null;
    dateEntered: string | null;
  }